Remove code to probe for 'saxon' script
The code that probed for the 'saxon' script did not actually make
sure that the script is the expected wrapper, even though the comment
claimed it is.  This makes xspec.sh fail when there is a 'saxon' script
in place that does not change the command line argument parsing behavior
like the expected 'saxon' script seems to.  I'm proposing to remove this
altogether and always use Saxon's native command line parsing routines
so that xspec.sh works on more systems.