Tail recursion in test:output-lines template
Put xsl:analyze-string inside a variable, tagging the unprocessed residue of the string separately as test:residue. Then, output the part of the string you processed (xsl:copy-of) and, lastly, call this template recursively if there is anything left to process.

Credit goes to Saxonica. This change works with XSLT 2.0. (The diff is noisy due to indentation.)

Beware: This change works with Saxon-EE 9.7.0.21J but not 9.8.0.12. I have reported the v9.8.0.12 problem to Saxonica.