Prefix "init" and "compile" in 'depends'
`init` and `compile` are a common target names.

`depends="init"` runs the `init` target in the importing build file, whereas `depends="xspec.init"` runs the `init` target in the `xspec` build file.  Similarly for `compile`.

Conceivably, every target name in a `depends` in XSpec's `build.xml` should be prefixed with the `xspec` build name, but `init` and `compile` are the two most likely to be accidentally overridden by an importing build file.