Update tests
Move contents of test/schematron-023.xspec into
test/end-to-end/cases/schematron-023.xspec
and delete test/schematron-023.xspec, as some of its test suites fail
on purpose and thus should not be present in folder test, but only be
tested end-to-end.

Update test/schematron/schut-to-xspec-012-out.xspec as
schut-to-xspec.xsl has been changed.