Substantive message when Saxon is new enough
Original code used only a space character with xsl:message
when the check for deprecated Saxon versions passes.
Bats testing did not detect the line, so many tests failed, such
as the "cd to bin dir ..." test case. My first workaround was a
nonbreaking space. That worked, but it would look strange
to users running XSpec tests in a win64 shell window. This
commit uses a substantive message, which is clearer.