feat(xslt): runtime error if context evaluates to empty sequence (#1780)
* feat(xslt): runtime error if context evaluates to empty sequence

Empty-sequence context is not productive and is
probably a user error. Issue an error message,
"Context is an empty sequence."

This change is for XSLT only. Schematron might benefit
from the same error checking in a subsequent pull request.

* test: make context nonempty in pre-existing tests

Where a test uses an empty context and the test
serves its purpose equally well with a nonempty context,
populate `x:context`.

* test: increase Windows bats test case count limit

* Use tabs when the rest of the file does

* test: don't expect error at end of line

win-bats test running on Azure has error message
followed immediately, on the same line, by the
path to the missing file.

* Centralize checking for empty context

* Add tests for context still empty after inheritance

* New bats test cases and updated messages

- New test cases are for inherited contexts that evaluate to empty sequence
- Message updates are due to XSLT change