build(oxygen): add-on v3.0.1
This v3.0.1 add-on has most of the changes we expect to
be in XSpec v3.0, but it excludes the XSLT Code Coverage
upgrade.

The next add-on version will include the
XSLT Code Coverage upgrade. Creating separate
add-on versions might make it easier for users
to compare the old and new functionality.