update end-to-end expected results
For the query/issue-355.* files, the report now renders
"pseudo-function" instead of "pseudo-other", as
a result of changing src/common/report-sequence.*
to use "instance of function(*)" instead of the
custom rep:instance-of-function function.

The items are function items, so "pseudo-function"
seems more appropriate. It also matches the XSLT
result. Related history: #870