In xsl:param test, include global not overridden
Mimics the difference that the Java change in #1918 makes
in test/global-override.xspec.

The overridden global xsl:param is marked as ignored, but an
open design question is whether it should instead be marked
as missed. The comment in this revision says "Expected miss"
in anticipation of changing the treatment of global params.