refactor(report): simplify XSLT computation of code coverage status (#1966)
* Refactor computation of XSLT code coverage status

Generated expected results of end-to-end coverage tests
are unchanged. Changing some of the rules for code coverage
status will be done in subsequent pull requests. The goal
of this refactoring is to make the rules more apparent,
so they will be easier to maintain and change if needed.

* Remove unneeded local:coverage function

All but one call to the function were already replaced
with apply-templates mode="coverage". Replace the
last remaining call and remove the function declaration.

* Simplify use of module-id

Retrieve module-id closer to where it's needed,
and reduce instances of passing it from one code
unit to the next.