fix(report): coverage for xsl:for-each and xsl:for-each-group (#1967)
* Refactor computation of XSLT code coverage status

Generated expected results of end-to-end coverage tests
are unchanged. Changing some of the rules for code coverage
status will be done in subsequent pull requests. The goal
of this refactoring is to make the rules more apparent,
so they will be easier to maintain and change if needed.

* Remove unneeded local:coverage function

All but one call to the function were already replaced
with apply-templates mode="coverage". Replace the
last remaining call and remove the function declaration.

* Simplify use of module-id

Retrieve module-id closer to where it's needed,
and reduce instances of passing it from one code
unit to the next.

* Add test for xsl:for-each-group w/ untraced child

* xsl:for-each(-group) to use trace data

Fixes #1949
Fixes #1950

* Regenerate expected results

The change of certain `xsl:for-each` and `xsl:for-each-group`
markings from "missed" to "hit" represents a bug fix for
the case where these elements' only children are untraced.