fix(report): coverage for top-level elements in stylesheet (#1968)
* Refactor computation of XSLT code coverage status

Generated expected results of end-to-end coverage tests
are unchanged. Changing some of the rules for code coverage
status will be done in subsequent pull requests. The goal
of this refactoring is to make the rules more apparent,
so they will be easier to maintain and change if needed.

* Remove unneeded local:coverage function

All but one call to the function were already replaced
with apply-templates mode="coverage". Replace the
last remaining call and remove the function declaration.

* Simplify use of module-id

Retrieve module-id closer to where it's needed,
and reduce instances of passing it from one code
unit to the next.

* Narrow scope for top-level non-XSLT elements

Side effect is to stop ignoring certain top-level XSLT elements

* Explicitly ignore top-level declarations

Many top-level XSLT declaration elements were marked "ignored" because
of the rule for top-level non-XSLT elements. Now that the latter rule is scoped
to non-XSLT elements, reimplement the logic that marks XSLT declarations
as ignored.

* Regenerate coverage test results

The change in xsl-param-01xsl coverage reflects a bug fix.
Global parameters that XSpec overrides are correctly
marked as "missed" rather than "ignored". They shouldn't
be blindly ignored, because global `xsl:param` elements
follow the Use Trace Data rule.