Updates based on last merge
- In test for xsl:where-populated, getLocation causes trace to
have a hit with correct column number for xsl:perform-sort,
which changes the coverage status to "hit" for that element
and its xsl:where-populated parent.
- In test for xsl:assert, the xsl:assert element on line 23 is marked
"unknown" due to #1986 (which anticipates no hits on xsl:assert
in Saxon 12.5 and sometimes wrong data in Saxon 12.4), even
though the trace has a correct hit on this element. That hit is why
xsl:on-completion on line 22 is marked as "hit". Leaving the line 22
comment in place for now, and it will become accurate in Saxon 12.5.