fix(report): coverage for variables, their descendants, and others (#1978)
* Remove special case for xsl:variable descendants

Changes seen in regenerated expected coverage reports:

- text-node: Known issue with text nodes being marked as missed
- xsl-evaluate: Known issues with xsl:evaluate not traced and text nodes marked as missed
- xsl-for-each-group: Known issue with xsl:map-entry not traced
- xsl-map: Known issue with xsl:map and xsl:map-entry not traced
- xsl-perform-sort: Known issue with xsl:sort child of xsl:perform-sort not traced
- xsl-sort: Known issue with xsl:sort child of xsl:perform-sort not traced
- xsl-variable: Lines that changed from hit to missed are all expected misses
- xsl-with-param: Known issue with xsl:evaluate and xsl:with-param not traced

* Adjust behavior for xsl:perform-sort and xsl:sort

- xsl:perform-sort: Use Child Data
- xsl:sort: Use Parent Status

* Adjust behavior for xsl:map and xsl:map-entry

Rule is "Unknown".

* Change xsl:map and xsl:map-entry to Use Child Data

If a sequence constructor has a node marked as hit,
the Use Child Data rule does better than Unknown.

* Adjust behavior for xsl:evaluate

Rule is "Use Child Data".

Also, enhance coverage test for xsl:evaluate to include
- xsl:evaluate element that is marked as 'hit' due to grandchild with hit in trace
- xsl:evaluate element that is unexecuted and marked as 'unknown'

* Adjust behavior for xsl:with-param

Rule is "Use Parent Status"

* Remove special case for global variables

If a global variable is not hit, it is missed. Checking the
coverage status of a following sibling of a global variable
is inconsistent with how top-level XSLT elements work.

* Fix expected result in test

* Regenerate expected results after merge

* Update expected results and comments after merge

- In xsl-where-populated-01.xsl, line 29 xsl:perform-sort is "unknown"
by Use Descendant Data rule because the only executable
descendant is xsl:sort, which is untraceable.
Line 30 xsl:sort is "unknown" by Use Parent Status rule because
xsl:perform-sort is "unknown".

- In xsl-with-param-01.xsl, line 46 xsl:evaluate is "unknown"
by Use Descendant Data rule because the element has an
executable, untraceable descendant node (xsl:with-param)
but no hits on any descendants. The child xsl:with-param nodes
are "unknown" by the Use Parent Status rule.

* Adjust behavior for xsl:sort descendants

- Use Parent Data for child elements
- Hit vs. Unknown for deeper descendants
- Update xslt-code-coverage-by-element.md
- Add test cases for children w/ and w/o hits in trace
- Add test cases for deeper descendants
- In xsl:evaluate test, add 'unknown' test case now that xsl:sort/xsl:evaluate is 'hit'

* Try to get same status with Saxon 12.4 and 12.5

Saxon 12.4 and 12.5 trace different descendants of xsl:map
in this test, but it is useful to have an example that shows
the Use Descendant Data rule taking effect.