fix(report): add getLocation method to trace listener (#1983)
* Add getLocation method to trace listener

See #1971.

Originally investigated by Adrian Bird (birdya22) with an
eye toward Saxon 12.5, this method appears to have some
good effects with Saxon 12.4, too, and no bad side effects
that I'm aware of.

Co-Authored-By: birdya22 <168069809+birdya22@users.noreply.github.com>

* Regenerate expected coverage results

with Saxon 12.4

* Updates based on last merge

- In test for xsl:where-populated, getLocation causes trace to
have a hit with correct column number for xsl:perform-sort,
which changes the coverage status to "hit" for that element
and its xsl:where-populated parent.
- In test for xsl:assert, the xsl:assert element on line 23 is marked
"unknown" due to #1986 (which anticipates no hits on xsl:assert
in Saxon 12.5 and sometimes wrong data in Saxon 12.4), even
though the trace has a correct hit on this element. That hit is why
xsl:on-completion on line 22 is marked as "hit". Leaving the line 22
comment in place for now, and it will become accurate in Saxon 12.5.

* Updates based on last merge

In merge with #1978, I picked 'master' branch copy of conflicted files.
This commit regenerates expected results, reflecting the effect of
getLocation.

* Update comment in xsl-on-non-empty-01.xsl

#1989 marked a particular xsl:sequence element as 'unknown'
but #1983 + Saxon 12.4 makes it 'hit', so update code comment in XSLT.

---------

Co-authored-by: birdya22 <168069809+birdya22@users.noreply.github.com>