fix(schematron): handle x:context with href and select
Without this change, `x:context[@href][@select]` throws
an error, `XPDY0002  The context item is absent`, pointing
to the compiled XSLT file.

With this change, XSpec constructs a temporary document
for the href case as in the case of embedded user content.

Note: This node selection operation should consider the
base URI to be undefined, like the embedded case but
unlike other node selection operations that use `@href`.