Comment and default for $report-theme
The comment includes the full filenames of the color-specific CSS
files so that if someone searches the repo for the filenames, they'll
get a result. In the executable code, the filenames are constructed
by concatenation.

Use 'default' as the default value and compute the effective
value afterward. That way, the Ant property, environment variable,
or Oxygen transformation scenario parameter can specify an
innocuous value to pick up the effective-default theme.