feat(report): black/white themes for XSpec reports (#2055)
* feat(report): black/white themes

* Ant property, xspec.html.report.theme

The xspec.html.report.theme Ant property provides
access to the report-theme XSLT parameter in format-utils.xsl.

* Env var, XSPEC_HTML_REPORT_THEME

The XSPEC_HTML_REPORT_THEME environment variable provides
access to the report-theme XSLT parameter in format-utils.xsl.

* Test report-theme interfaces (Ant/CLI)

* Expose xspec.html.report.theme in Oxygen

Add the xspec.html.report.theme Ant property to the
Oxygen Ant transformation scenarios for test results
and XSLT code coverage.

Without this change, users would have to set something like
-Dxspec.html.report.theme=whiteblack
in the Additional Ant arguments field on the Options tab.

* Comment and default for $report-theme

The comment includes the full filenames of the color-specific CSS
files so that if someone searches the repo for the filenames, they'll
get a result. In the executable code, the filenames are constructed
by concatenation.

Use 'default' as the default value and compute the effective
value afterward. That way, the Ant property, environment variable,
or Oxygen transformation scenario parameter can specify an
innocuous value to pick up the effective-default theme.

* Make the default black text on white background

* Adapt test labels for blackwhite theme

But also retain color names referring to classic theme,
in case that helps test maintainers.

* Regenerate expected results with blackwhite theme