fix(cli): saxon script is fallback, not first try
The SAXON_CP and SAXON_HOME environment variables are
the cross-platform documented ways to indicate where to
find Saxon. The "saxon" script from the EXPath Packaging System
is a tested but undocumented alternative that has been
implemented only on Linux and macOS. The implementation
interferes with other ways a user might have a "saxon"
script on the path, so we should give SAXON_CP and
SAXON_HOME higher precedence in the logic.