fix(cli): remove problematic support for saxon script (#2073)
* fix(cli): saxon script is fallback, not first try

The SAXON_CP and SAXON_HOME environment variables are
the cross-platform documented ways to indicate where to
find Saxon. The "saxon" script from the EXPath Packaging System
is a tested but undocumented alternative that has been
implemented only on Linux and macOS. The implementation
interferes with other ways a user might have a "saxon"
script on the path, so we should give SAXON_CP and
SAXON_HOME higher precedence in the logic.

* Test that SAXON_CP has precedence over saxon

Test for #2072.

The win-bats/collection.xml change is just to
keep the two test sets aligned. The Windows
script to run XSpec does not support falling back
to a "saxon" script on the path.

* Drop support for saxon script