build(oxygen): add-on v3.2.1, w/ automatic Git tag for XSpec version (#2076)
* GitHub action to push tag based on VERSION

Uses anothrNick/github-tag-action but the logic
is much simpler: whenever the `VERSION` file is
pushed to the `master` branch, use that file's content
as the version number, prefix it with `v`, and push
a tag with that name to the remote. The remote
could be xspec/xspec or a fork of it.

* Oxygen add-on for v3.2.1

* Disable add-on when using this project file

The part between "4/" and "/xspec.framework" is the
directory name at the end of the add-on installation at
...\com.oxygenxml\extensions\v27.0\frameworks\io.xspec.xspec\xspec-3.2.1

* Mention 2078 now that it has been merged

Also, remove the item about change in release automation.

* Use v3.2.1 as the release candidate for v3.2