fix: reject x:call that omits what to call
If x:call (after resolving any inheritance) does not have either a
`function` or `template` attribute, there is nothing to call.
This commit adds a user-facing error message for that situation.

Without this change, the situation leads to a user-unfriendly
error from the compiler:
```
src\compiler\xslt\compile\compile-scenario.xsl:198:34: Fatal Error! Conditional expression: None of the conditions is satisfied, so an empty sequence is returned, but this is not allowed as the value of variable $invocation-type
```