ci: bump Java from 8 to 11 on Azure Pipelines (#2164)
* ci: bump Java from 8 to 11 on macOS Azure

* Adapt Java version check for BaseX/Java11

When we drop Java 8 testing entirely, we can remove
the code that looks for " 1.8".

* Don't compile Java on Azure

It will be done in GitHub Actions, and the
last-git-status.cmd script fails if Azure compiles
on Java 11 while our .class files were compiled
with Java 8.

This change is only until all our CI uses Java 11
instead of Java 8.