ci: bump Java from 8 and 11 to 17 and 21 (#2201)
* Test with Java 17 and 21

* Compile with Java 17

* Remove version condition for BaseX

Testing with Java 17 and 21 means BaseX is
always available.

* Azure macOS test should use Java 17

Had been 8, which was the older between 8 and 11.
Was bumped to 11 temporarily in #2164.
Now, use 17 as the older of the two Java versions the
repo will test with.

* AppVeyor should use Java 17

The default image for Windows is Visual Studio 2015,
which doesn't have Java 17. Select Visual Studio 2019 instead.

References:
https://www.appveyor.com/docs/windows-images-software/#java
https://www.appveyor.com/docs/appveyor-yml/

* Skip Java compilation/check on AppVeyor